# Conflicts:
#	shop-dao/src/main/java/com/masiis/shop/dao/mall/user/SfUserRelationMapper.java
#	shop-service/src/main/java/com/masiis/shop/web/mall/service/user/SfUserRelationService.java

+++ shop-dao/src/main/java/com/masiis/shop/dao/mall/message/SfMessageSrRelationMapper.java
@@ -26,5 +26,5 @@ public interface SfMessageSrRelationMapper {
 
     int updateByPrimaryKey(SfMessageSrRelation record);
 
-    Integer countNumsFromUser(@Param("userId") Long userId);
+    Long countNumsFromUser(@Param("userId") Long userId);
 }
